Share authentication logic with Admin namespaced controllers
