+++ app/controllers/admin/application_controller.rb
@@ -3,5 +3,6 @@
 # The Admin CMS features use erb views and require access to
 # the protect_from_forgery methods provided by ActionController::Base
 class Admin::ApplicationController < ActionController::Base
+  include Authentication
   layout "cms"
 end
