+++ app/controllers/admin/cms/collections_controller.rb
@@ -1,4 +1,5 @@
 class Admin::Cms::CollectionsController < Admin::ApplicationController
+  before_filter :authenticate_admin!
   before_action :set_collection, only: [:show, :edit, :update]
 
   def show
