+++ app/controllers/admin/cms_controller.rb
@@ -1,4 +1,6 @@
 class Admin::CmsController < Admin::ApplicationController
+  before_filter :authenticate_admin!
+  
   def show
     @collection = Collection.all
   end
